refactor: rename typedata package to typeddata and update references
