+++ paymaster/build_txn.go
@@ -8,7 +8,7 @@ import (
 
 	"github.com/NethermindEth/juno/core/felt"
 	"github.com/NethermindEth/starknet.go/client/rpcerr"
-	"github.com/NethermindEth/starknet.go/typedata"
+	"github.com/NethermindEth/starknet.go/typeddata"
 )
 
 // BuildTransaction receives the transaction the user wants to execute. Returns the typed
@@ -389,7 +389,7 @@ type BuildTransactionResponse struct {
 	Parameters *UserParameters `json:"parameters"`
 	// The typed data for for `invoke` and `deploy_and_invoke` transaction types.
 	// It's `nil` for `deploy` transaction types.
-	TypedData *typedata.TypedData `json:"typed_data,omitempty"`
+	TypedData *typeddata.TypedData `json:"typed_data,omitempty"`
 	// The fee estimation for the transaction
 	Fee *FeeEstimate `json:"fee"`
 }
