+++ typeddata/testData/example_enum.json
