+++ typeddata/testData/mail_StructArray.json
